don't create symlink if already exists

+++ install.sh
@@ -5,4 +5,4 @@ readonly SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" &> /dev/null && pw
 [ ! -d venv ] && python3 -m venv "$SCRIPT_DIR/venv"
 source "$SCRIPT_DIR/venv/bin/activate"
 pip3 install -r "$SCRIPT_DIR/requirements.txt"
-ln -s "$SCRIPT_DIR/brag" ~/.local/bin/brag
+[ ! -L ~/.local/bin/brag ] && ln -s "$SCRIPT_DIR/brag" ~/.local/bin/brag
